Single Port RAM
Single-port Random Access Memory (RAM) is a type of memory that allows for the reading and writing of data through a single port or interface.
single-port RAM has only one data port, which means read and write operations cannot occur simultaneously at different addresses. If a write operation is in progress, a read operation must wait, and vice versa.