CARRY LOOK AHEAD ADDER
It improves the speed of addition by reducing the propagation delay unlike the ripple carry adders. In a ripple carry adder, the carry output from one stage ripples through to the next stage, causing a delay in the computation of the final result.
A Carry Look-Ahead Adder minimizes this delay by calculating the carry outputs for each stage in advance, hence it is resulting in faster addition.