FSM
A FSM or finite state machine, is a circuit that contains a pre-defined number of states. The machine can exist in one and only one state at a time. The circuit transitions between states based on a triggering event, most commonly the edge of a clock, in addition to the values of any inputs of the machine. The number of states and all possible transitions are pre-defined. Through the use of states and a pre-defined sequence of transitions, the circuit is able to make decisions on the next state to transition to based on a history of past states. This allows the circuit to create outputs that are more intelligent compared to a simple combinational logic circuit that has outputs based only on the current values of the inputs.
There are two different types of output conditions for a state machine:-
i. When the output only depends on the current state of the machine. This type of system is called a Moore Machine.
ii. The second output condition is when the outputs depend on both the current state and the system inputs. This type of system is called a Mealy Machine.