SPI
Serial Peripheral Interface (SPI) is a synchronous serial communication interface used for short-distance communication between microcontrollers, sensors, and other devices.
SPI is a full-duplex interface, meaning that data can be transmitted and received simultaneously.
SPI requires a minimum of four pins: MOSI (Master Out Slave In), MISO (Master In Slave Out), SCK (Serial Clock), and SS (Slave Select).
MOSI is used to transmit data from the master to the slave, while MISO is used to transmit data from the slave to the master.
SCK is the clock signal generated by the master, and SS is used to select the slave device that will communicate with the master.